requirements.txt
fastapi==0.128.7
uvicorn==0.40.0
requests==2.32.5
python-dotenv==1.2.1
numpy==2.4.2
cachetools==7.0.1
pandas==3.0.0
openpyxl==3.1.5
pydantic==2.12.5
